AuthContext を contexts ディレクトリに移動
